// src/app/components/shared/interfaces.ts
export interface ICustomer {
    id: number;
    name: string;
    city: string;
    orderTotal?: number;
    customerSince: any;
}

export interface IOrder {
    customerId: number;
    orderItems: IOrderItem[];
}

export interface IOrderItem {
    id: number;
    productName: string;
    itemCost: number;
}



// interface is a code contract - if you use ICustomer for example you have to use the items from the customer definition